Implement Display and Debug using LowerHex
`Display` and `LowerHex` offer an 'alternate' option, hex
representations by convention should include `0x` if the alternate
format is printed.

Currently we implement `Display`, `Debug`, and `LowerHex` all using the
same macro. This has two problems, the biggest of which is that we are
not correctly handling the alternate format. The second issue is code
duplication.

Modify the `impl_hex_fmt` macro to implement `LowerHex` (incl. alternate
form), then implement `Display` and `Debug` in the macro making calls to
`LowerHex`.